_Legacy_Flet/requirements.txt
flet
msal
supabase
python-dotenv
requests
pandas
openpyxl